// src/interfaces/jmt.interface.ts
export interface coverMinimap {
	mapImage: string
	rect: number
	show: boolean
	position: {
		x: number
		y: number
	}
}